feat: Rename the rule to `no-mutable-array`
