+++ tests/rules/assertions/noMutableArray.js
