+++ tests/rules/index.js
@@ -16,13 +16,13 @@ const reportingRules = [
   'delimiter-dangle',
   'generic-spacing',
   'no-dupe-keys',
+  'no-mutable-array',
   'no-primitive-constructor-types',
   'no-types-missing-file-annotation',
   'no-unused-expressions',
   'no-weak-types',
   'object-type-delimiter',
   'require-parameter-type',
-  'require-read-only-array',
   'require-return-type',
   'require-valid-file-annotation',
   'require-variable-type',
